unsafe builtin blocks support
