+++ web/src/beta/lib/core/Crust/Infobox/Block/builtin/Image/index.stories.tsx
