+++ web/src/beta/lib/core/Crust/Infobox/Block/builtin/Image/index.tsx
@@ -3,8 +3,8 @@ import React, { useCallback, useState } from "react";
 import Icon from "@reearth/beta/components/Icon";
 import { styled } from "@reearth/services/theme";
 
-import { CommonProps as BlockProps } from "..";
-import { Border, Title } from "../utils";
+import { CommonProps as BlockProps } from "../..";
+import { Border, Title } from "../../utils";
 
 export type Props = BlockProps<Property>;
