+++ web/src/beta/lib/core/Crust/Infobox/Block/builtin/Video/index.tsx
@@ -4,8 +4,8 @@ import Player from "react-player";
 import Icon from "@reearth/beta/components/Icon";
 import { styled } from "@reearth/services/theme";
 
-import { CommonProps as BlockProps } from "..";
-import { Border, Title } from "../utils";
+import { CommonProps as BlockProps } from "../..";
+import { Border, Title } from "../../utils";
 
 export type Props = BlockProps<Property>;
