+++ web/src/beta/lib/core/Crust/types.ts
@@ -12,3 +12,21 @@ export type {
 } from "../Map";
 export type { Theme } from "./theme";
 export type { InteractionModeType } from "./interactionMode";
+
+export type UnsafeBuiltinPlugin = {
+  id: string;
+  name: string;
+  widgets: UnsafeBuiltinWidget[];
+  blocks: UnsafeBuiltinBlock[];
+};
+
+type UnsafeBuiltinWidget = UnsafeBuiltinPluginExtension<"widget">;
+
+type UnsafeBuiltinBlock = UnsafeBuiltinPluginExtension<"block">;
+
+type UnsafeBuiltinPluginExtension<T extends "widget" | "block"> = {
+  type: T;
+  extensionId: string;
+  name: string;
+  component: React.FC;
+};
